fix text

+++ files/routes/awards.py
@@ -188,7 +188,7 @@ def award_thing(v, thing_type, id):
 			send_repeatable_notification(v.id, msg)
 			author = v
 		elif kind != 'spider':
-			awarded_coins = AWARDS[kind]['price'] * COSMETIC_AWARD_COIN_AWARD_PCT if AWARDS[kind]['cosmetic'] else 0
+			awarded_coins = int(AWARDS[kind]['price'] * COSMETIC_AWARD_COIN_AWARD_PCT) if AWARDS[kind]['cosmetic'] else 0
 			if AWARDS[kind]['cosmetic']:
 				author.coins += awarded_coins
